Refactoring to separate core/gui/personality.
