+++ core/recognizer.py
