--- etc/greeting
@@ -1,1 +1,0 @@
-hey there
