--- etc/name
@@ -1,1 +1,0 @@
-vav...
