--- etc/question
@@ -1,1 +1,0 @@
-how are you?
