+++ gui/from_assistant.txt
@@ -1,0 +1,39 @@
+FROM __init__
+
+    #if self.options['interface']:
+    #    if self.options['interface'] == "g":
+    #        from bin.gui import GTKInterface as UI
+    #    elif self.options['interface'] == "gt":
+    #        from bin.gui import GTKTrayInterface as UI
+    #    else:
+    #        print("no GUI defined")
+    #        sys.exit()
+
+    #    self.ui = UI(self.options, self.options['continuous'])
+    #    self.ui.connect("command", self.process_command)
+    #    # Able To Load Icon?
+    #    icon = self.load_resource("icon_small.png")
+    #    if icon:
+    #        self.ui.set_icon_active_asset(icon)
+    #    # Able To Load Inactive Icon?
+    #    icon_inactive = self.load_resource("icon_inactive_small.png")
+    #    if icon_inactive:
+    #        self.ui.set_icon_inactive_asset(icon_inactive)
+
+
+FROM recognizer_finished
+
+    # If There Is A Ui And We Are Not In Continuous Listen
+    #if self.ui:
+    #    if not self.continuous_listen:
+    #        # Stop Listening
+    #        self.recognizer.pause()
+    #    # Tell Ui To Finish
+    #    self.ui.finished(t)
+
+
+FROM run
+
+    #if self.ui:
+    #    self.ui.run()
+    #else:
