+++ personalities/default/cache/assisthistory
@@ -1,0 +1,9 @@
+HELLO
+HEY
+HELLO
+TRUE
+WHAT HAPPENED?
+THANK
+HOME
+HOME
+HEY
