+++ personalities/default/cache/hash.json
@@ -1,0 +1,1 @@
+{"language": "b1c6ecf86edf96894c037f907a3f1c5fb5412457b81df81d7ff363e403b17d9f", "voice_commands": "e84db0de153f0e7f21dbedb3d2ccc5124ad2098045129e5ffd1d91eccfcc42f3"}
