+++ personalities/default/commands/nettest.py
