+++ personalities/default/commands/newsworld.py
