+++ personalities/default/docs/jokes.txt
