+++ personalities/default/docs/neuromancer.txt
