+++ personalities/default/docs/projectm_stella.ppl
