+++ personalities/default/etc/thanks
