+++ personalities/default/language/sentences.corpus
